requirements.txt
bcrypt==3.2.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
Flask==2.0.1
Flask-Bcrypt==0.7.1
flask-crontab==0.1.2
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
nulltype==2.3.1
pickleshare==0.7.5
plaid-python==8.0.0b12
prompt-toolkit==3.0.18
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.16.0
SQLAlchemy==1.4.22
twilio==6.63.1
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
WTForms==2.3.3